+++ src/components/FormEdit.js
@@ -27,7 +27,7 @@ export default class extends Component {
   }
 
   static getDerivedStateFromProps(nextProps, prevState) {
-    if (nextProps.form && prevState.form._id !== nextProps.form._id) {
+    if (nextProps.form && (prevState.form._id !== nextProps.form._id || prevState.form.modified !== nextProps.form.modified)) {
       return {
         form: _cloneDeep(nextProps.form)
       };
@@ -62,12 +62,11 @@ export default class extends Component {
 
   render() {
     const {form} = this.state;
-    const {saveText, title} = this.props;
+    const {saveText} = this.props;
+    console.log(form.modified);
 
     return (
       <div>
-        <h2>{title}</h2>
-        <hr />
         <div className="row">
           <div className="col-lg-2 col-md-4 col-sm-4">
             <div id="form-group-title" className="form-group">