+++ src/components/index.js
@@ -1,3 +1,4 @@
+export {default as Errors} from './Errors';
 export {default as Form} from './Form';
 export {default as FormBuilder} from './FormBuilder';
 export {default as FormEdit} from './FormEdit';